Version 2.3.6.0 of the AWS .NET SDK

+++ AWSSDK_DotNet.IntegrationTests/Tests/Route53.cs
@@ -6,6 +6,8 @@ using System.Text;
 
 using Amazon.Route53;
 using Amazon.Route53.Model;
+using System.Threading;
+using Amazon;
 
 namespace AWSSDK_DotNet.IntegrationTests.Tests
 {
@@ -14,7 +16,9 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
     {
         private const string COMMENT = "comment";
         private const string ZONE_NAME = "aws.sdk.com.";
-        private static readonly string CALLER_REFERENCE = Guid.NewGuid().ToString();
+        private static string CALLER_REFERENCE { get { return Guid.NewGuid().ToString(); } }
+        private static TimeSpan pollingPeriod = TimeSpan.FromSeconds(30);
+        private static TimeSpan maxWaitTime = TimeSpan.FromMinutes(5);
 
         // The ID of the zone we created in this test
         private string createdZoneId;
@@ -32,20 +36,27 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
         [TestCleanup()]
         public void TearDown()
         {
-            try
+            if (!string.IsNullOrEmpty(createdZoneId))
             {
-                Client.DeleteHostedZone(new DeleteHostedZoneRequest { Id = createdZoneId });
+                try
+                {
+                    Client.DeleteHostedZone(new DeleteHostedZoneRequest { Id = createdZoneId });
+                }
+                catch { }
             }
-            catch (Exception) { }
         }
 
         
-        // Runs through each of the APIs in the Route 53 client to make sure we can
-        // correct send requests and unmarshall responses.         
+        // Runs through a number of the APIs in the Route 53 client to make sure we can
+        // correct send requests and unmarshall responses.
         [TestMethod]
         [TestCategory("Route53")]
         public void TestRoute53()
         {
+            var geoLocations = Client.ListGeoLocations(new ListGeoLocationsRequest()).GeoLocationDetailsList;
+            Assert.IsNotNull(geoLocations);
+            Assert.AreNotEqual(0, geoLocations.Count);
+
             CreateHostedZoneRequest createRequest = new CreateHostedZoneRequest
             {
                 Name = ZONE_NAME,
@@ -72,7 +83,7 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
 
 
             // List Hosted Zones
-            List<HostedZone> hostedZones = Client.ListHostedZones(new ListHostedZonesRequest()).HostedZones;
+            List<HostedZone> hostedZones = Client.ListHostedZones().HostedZones;
             Assert.IsTrue(hostedZones.Count > 0);
             foreach (HostedZone hostedZone in hostedZones)
             {
@@ -109,7 +120,8 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
                 Name = ZONE_NAME,
                 ResourceRecords = existingResourceRecordSet.ResourceRecords,
                 TTL = existingResourceRecordSet.TTL + 100,
-                Type = existingResourceRecordSet.Type
+                Type = existingResourceRecordSet.Type,
+                HealthCheckId = null
             };
 
             changeInfo = Client.ChangeResourceRecordSets(new ChangeResourceRecordSetsRequest
@@ -127,7 +139,6 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
             }).ChangeInfo;
 
             assertValidChangeInfo(changeInfo);
-
 
             // Delete Hosted Zone
             DeleteHostedZoneResult deleteHostedZoneResult = Client.DeleteHostedZone(new DeleteHostedZoneRequest { Id = createdZoneId });
@@ -154,29 +165,279 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
             Assert.IsNotNull(createResponse.HealthCheck.Id);
             Assert.AreEqual(10, createResponse.HealthCheck.HealthCheckConfig.RequestInterval);
             Assert.AreEqual(5, createResponse.HealthCheck.HealthCheckConfig.FailureThreshold);
+            string healthCheckId = createResponse.HealthCheck.Id;
 
             var listResponse = Client.ListHealthChecks();
-            Assert.IsNotNull(listResponse.HealthChecks.FirstOrDefault(x => x.Id == createResponse.HealthCheck.Id));
+            Assert.IsNotNull(listResponse.HealthChecks.FirstOrDefault(x => x.Id == healthCheckId));
 
-            Client.DeleteHealthCheck(new DeleteHealthCheckRequest() { HealthCheckId = createResponse.HealthCheck.Id });
+            GetHealthCheckStatusResponse status = null;
+            var stopTime = DateTime.Now + maxWaitTime;
+
+            while (DateTime.Now < stopTime)
+            {
+                try
+                {
+                    status = Client.GetHealthCheckStatus(new GetHealthCheckStatusRequest
+                    {
+                        HealthCheckId = healthCheckId
+                    });
+                    break;
+                }
+                catch (NoSuchHealthCheckException)
+                {
+                    Thread.Sleep(TimeSpan.FromSeconds(10));
+                }
+            }
+            Assert.IsNotNull(status);
+            Assert.IsNotNull(status.HealthCheckObservations);
+
+            var healthCheck = Client.GetHealthCheck(new GetHealthCheckRequest
+            {
+                HealthCheckId = healthCheckId
+            }).HealthCheck;
+            Assert.IsNotNull(healthCheck);
+            Assert.IsNotNull(healthCheck.Id);
+            Assert.IsNotNull(healthCheck.HealthCheckConfig);
+
+            var tagSet = Client.ListTagsForResource(new ListTagsForResourceRequest
+            {
+                ResourceType = TagResourceType.Healthcheck,
+                ResourceId = healthCheckId
+            }).ResourceTagSet;
+            Assert.IsNotNull(tagSet);
+            Assert.IsNotNull(tagSet.ResourceId);
+            Assert.IsNotNull(tagSet.ResourceType);
+            Assert.IsNotNull(tagSet.Tags);
+            Assert.AreEqual(0, tagSet.Tags.Count);
+
+            Client.ChangeTagsForResource(new ChangeTagsForResourceRequest
+            {
+                ResourceType = TagResourceType.Healthcheck,
+                ResourceId = healthCheckId,
+                AddTags = new List<Tag>
+                {
+                    new Tag { Key = "Test", Value = "true" }
+                }
+            });
+
+            tagSet = Client.ListTagsForResource(new ListTagsForResourceRequest
+            {
+                ResourceType = TagResourceType.Healthcheck,
+                ResourceId = healthCheckId
+            }).ResourceTagSet;
+            Assert.IsNotNull(tagSet);
+            Assert.IsNotNull(tagSet.ResourceId);
+            Assert.IsNotNull(tagSet.ResourceType);
+            Assert.IsNotNull(tagSet.Tags);
+            Assert.AreEqual(1, tagSet.Tags.Count);
+            Assert.AreEqual("Test", tagSet.Tags[0].Key);
+            Assert.AreEqual("true", tagSet.Tags[0].Value);
+
+            Client.DeleteHealthCheck(new DeleteHealthCheckRequest() { HealthCheckId = healthCheckId });
 
             listResponse = Client.ListHealthChecks();
-            Assert.IsNull(listResponse.HealthChecks.FirstOrDefault(x => x.Id == createResponse.HealthCheck.Id));
+            Assert.IsNull(listResponse.HealthChecks.FirstOrDefault(x => x.Id == healthCheckId));
+
+
         }
 
+        [TestMethod]
+        [TestCategory("Route53")]
+        public void DelegationSetTests()
+        {
+            List<string> createdSets = new List<string>();
 
-        
+            var sets = Client.ListReusableDelegationSets(new ListReusableDelegationSetsRequest());
+            var setCount = sets.DelegationSets.Count;
+
+            var callerReference = "DNSMigration" + DateTime.Now.ToFileTime();
+            var createResponse = Client.CreateReusableDelegationSet(new CreateReusableDelegationSetRequest
+            {
+                CallerReference = callerReference
+            });
+            Assert.IsNotNull(createResponse.Location);
+            var delegationSet = createResponse.DelegationSet;
+            Assert.IsNotNull(delegationSet);
+            Assert.IsNotNull(delegationSet.CallerReference);
+            Assert.IsNotNull(delegationSet.Id);
+            Assert.IsNotNull(delegationSet.NameServers);
+            Assert.AreNotEqual(0, delegationSet.NameServers.Count);
+            createdSets.Add(delegationSet.Id);
+
+            sets = Client.ListReusableDelegationSets(new ListReusableDelegationSetsRequest());
+            Assert.AreEqual(setCount + 1, sets.DelegationSets.Count);
+
+            CreateHostedZoneRequest createRequest = new CreateHostedZoneRequest
+            {
+                Name = ZONE_NAME,
+                CallerReference = CALLER_REFERENCE,
+                HostedZoneConfig = new HostedZoneConfig { Comment = COMMENT },
+                DelegationSetId = delegationSet.Id
+            };
+            createdZoneId = Client.CreateHostedZone(createRequest).HostedZone.Id;
+
+            var hostedZoneInfo = Client.GetHostedZone(new GetHostedZoneRequest
+            {
+                Id = createdZoneId
+            });
+            Assert.IsNotNull(hostedZoneInfo.VPCs);
+            Assert.IsFalse(hostedZoneInfo.HostedZone.Config.PrivateZone);
+            Assert.AreEqual(delegationSet.Id, hostedZoneInfo.DelegationSet.Id);
+
+            var hostedZones = Client.ListHostedZones(new ListHostedZonesRequest
+            {
+                DelegationSetId = delegationSet.Id
+            }).HostedZones;
+            Assert.AreEqual(1, hostedZones.Count);
+
+            // add a second set
+            callerReference = "DNSMigration" + DateTime.Now.ToFileTime();
+            createResponse = Client.CreateReusableDelegationSet(new CreateReusableDelegationSetRequest
+            {
+                CallerReference = callerReference
+            });
+            delegationSet = createResponse.DelegationSet;
+            createdSets.Add(delegationSet.Id);
+
+            int totalSetCount = 0;
+            string nextMarker = null;
+            do
+            {
+                var response = Client.ListReusableDelegationSets(new ListReusableDelegationSetsRequest
+                {
+                    MaxItems = "1",
+                    Marker = nextMarker
+                });
+                totalSetCount += response.DelegationSets.Count;
+                nextMarker = response.NextMarker;
+            } while (!string.IsNullOrEmpty(nextMarker));
+            Assert.AreEqual(setCount + 2, totalSetCount);
+
+            Client.DeleteHostedZone(new DeleteHostedZoneRequest
+            {
+                Id = createdZoneId
+            });
+            createdZoneId = null;
+
+            foreach (var setId in createdSets)
+            {
+                Client.DeleteReusableDelegationSet(new DeleteReusableDelegationSetRequest
+                {
+                    Id = setId
+                });
+            }
+
+            sets = Client.ListReusableDelegationSets(new ListReusableDelegationSetsRequest());
+            Assert.AreEqual(setCount, sets.DelegationSets.Count);
+        }
+
+        [TestMethod]
+        [TestCategory("Route53")]
+        public void VPCTests()
+        {
+            var vpc1 = CreateVPC();
+
+            CreateHostedZoneRequest createRequest = new CreateHostedZoneRequest
+            {
+                Name = ZONE_NAME,
+                CallerReference = CALLER_REFERENCE,
+                HostedZoneConfig = new HostedZoneConfig { Comment = COMMENT },
+                VPC = vpc1
+            };
+            createdZoneId = Client.CreateHostedZone(createRequest).HostedZone.Id;
+
+            var hostedZoneInfo = Client.GetHostedZone(new GetHostedZoneRequest
+            {
+                Id = createdZoneId
+            });
+            Assert.IsNotNull(hostedZoneInfo.VPCs);
+            Assert.AreEqual(1, hostedZoneInfo.VPCs.Count);
+            Assert.IsTrue(hostedZoneInfo.HostedZone.Config.PrivateZone);
+
+            var vpc2 = CreateVPC();
+            var changeInfo = Client.AssociateVPCWithHostedZone(new AssociateVPCWithHostedZoneRequest
+            {
+                VPC = vpc2,
+                Comment = COMMENT,
+                HostedZoneId = createdZoneId
+            }).ChangeInfo;
+            Assert.IsNotNull(changeInfo);
+            Assert.IsNotNull(changeInfo.Comment);
+            assertValidChangeInfo(changeInfo);
+
+            hostedZoneInfo = Client.GetHostedZone(new GetHostedZoneRequest
+            {
+                Id = createdZoneId
+            });
+            Assert.IsNotNull(hostedZoneInfo.VPCs);
+            Assert.AreEqual(2, hostedZoneInfo.VPCs.Count);
+
+            changeInfo = Client.DisassociateVPCFromHostedZone(new DisassociateVPCFromHostedZoneRequest
+            {
+                HostedZoneId = createdZoneId,
+                VPC = vpc2
+            }).ChangeInfo;
+            assertValidChangeInfo(changeInfo);
+
+            hostedZoneInfo = Client.GetHostedZone(new GetHostedZoneRequest
+            {
+                Id = createdZoneId
+            });
+            Assert.IsNotNull(hostedZoneInfo.VPCs);
+            Assert.AreEqual(1, hostedZoneInfo.VPCs.Count);
+
+            changeInfo = Client.DeleteHostedZone(new DeleteHostedZoneRequest
+            {
+                Id = createdZoneId
+            }).ChangeInfo;
+            assertValidChangeInfo(changeInfo);
+
+            DeleteVPC(vpc1);
+            DeleteVPC(vpc2);
+        }
+
+        // Deletes a VPC
+        private void DeleteVPC(VPC vpc)
+        {
+            using (var ec2 = new Amazon.EC2.AmazonEC2Client())
+            {
+                ec2.DeleteVpc(new Amazon.EC2.Model.DeleteVpcRequest
+                {
+                    VpcId = vpc.VPCId
+                });
+            }
+        }
+
+        // Creates a VPC
+        private VPC CreateVPC()
+        {
+            var region = VPCRegion.FindValue(AWSConfigs.RegionEndpoint.SystemName);
+
+            using(var ec2 = new Amazon.EC2.AmazonEC2Client())
+            {
+                var ec2Vpc = ec2.CreateVpc(new Amazon.EC2.Model.CreateVpcRequest
+                {
+                    CidrBlock = "10.0.0.0/16",
+                    InstanceTenancy = Amazon.EC2.Tenancy.Default
+                }).Vpc;
+
+                return new VPC
+                {
+                    VPCRegion = region,
+                    VPCId = ec2Vpc.VpcId
+                };
+            }
+        }
+
         // Asserts that the specified HostedZone is valid and represents the same
         // HostedZone that we initially created at the very start of this test.        
         private void assertValidCreatedHostedZone(HostedZone hostedZone)
         {
-            Assert.AreEqual(CALLER_REFERENCE, hostedZone.CallerReference);
             Assert.AreEqual(ZONE_NAME, hostedZone.Name);
             Assert.IsNotNull(hostedZone.Id);
             Assert.AreEqual(COMMENT, hostedZone.Config.Comment);
         }
 
-        
         // Asserts that the specified DelegationSet is valid.                 
         private void assertValidDelegationSet(DelegationSet delegationSet)
         {
@@ -193,6 +454,12 @@ namespace AWSSDK_DotNet.IntegrationTests.Tests
             Assert.IsNotNull(change.Id);
             Assert.IsNotNull(change.Status);
             Assert.IsNotNull(change.SubmittedAt);
+
+            ChangeInfo retrievedChange = Client.GetChange(new GetChangeRequest { Id = change.Id }).ChangeInfo;
+            Assert.IsNotNull(retrievedChange);
+            Assert.IsNotNull(retrievedChange.Id);
+            Assert.IsNotNull(retrievedChange.Status);
+            Assert.IsNotNull(retrievedChange.SubmittedAt);
         }
     }
 }
